feat:page.tsx以外を_componentsに配置
