+++ src/app/page.tsx
@@ -1,7 +1,84 @@
 "use client";
 
-import Posts from "@/_components/Posts";
-// 最初に表示されるページ
+import { API_BASE_URL } from "@/_constants/constants";
+import { Post } from "@/_types/Post";
+import Link from "next/link";
+import { useEffect, useState } from "react";
+
 export default function Home() {
-  return <Posts />;
+  // 初期値[空配列]設定でpostsがundefindでmapメソッドのエラー回避
+  const [posts, setPosts] = useState<Post[]>([]);
+
+  // Loadingの格納・表示
+  // 初期ロード時は無効
+  const [loading, setLoading] = useState<boolean>(false);
+
+  // featchエラーの格納・表示
+  // JSで空配列[]はtruthy。nullはfalsy。
+  const [error, setError] = useState<Error | null>(null);
+
+  useEffect(() => {
+    const fetcher = async (): Promise<void> => {
+      try {
+        // ローディングインジケーター表示
+        setLoading(true);
+        const res = await fetch(`${API_BASE_URL}/posts`);
+        if (!res.ok) {
+          // consoleに表示
+          throw new Error("データが見つかりません");
+        }
+        const { posts } = (await res.json()) as { posts: Post[] };
+        setPosts(posts);
+      } catch (error) {
+        // errorは最初にunknown型として扱われる
+        if (error instanceof Error) {
+          setError(error);
+        }
+      } finally {
+        setLoading(false);
+      }
+    };
+    fetcher();
+  }, []);
+
+  // ローディング中
+  if (loading) {
+    return <div>読込中・・・</div>;
+  }
+
+  // fetchエラー
+  if (error) {
+    return <div> 記事取得エラー: {error.message}</div>;
+  }
+
+  // 取得したデータが空である場合
+  if (!loading && !posts) {
+    return <div>記事が存在しません</div>;
+  }
+
+  return (
+    <main>
+      <ul>
+        {/* 取得したpostsデータが配列なのでmapでloop */}
+        {posts.map((post) => (
+          <li key={post.id}>
+            <Link href={`posts/${post.id}`}>
+              <div className="post">
+                <div className="post-info">
+                  <p>{new Date(post.createdAt).toLocaleDateString()}</p>
+                  <ul>
+                    {post.categories.map((category) => (
+                      <li key={category}>{category}</li>
+                    ))}
+                  </ul>
+                </div>
+                <h1>{post.title}</h1>
+                <p dangerouslySetInnerHTML={{ __html: post.content }} />
+              </div>
+            </Link>
+          </li>
+        ))}
+      </ul>
+    </main>
+  );
 }
