+++ src/app/posts/[id]/page.tsx
@@ -1,12 +1,65 @@
 "use client";
 
-import Post from "@/_components/Post";
+import { API_BASE_URL } from "@/_constants/constants";
+import type { Post } from "@/_types/Post";
 import { useParams } from "next/navigation";
+import { useEffect, useState } from "react";
 
 export default function Page() {
-  const { id } = useParams<{ id: string }>();
+  const { id } = useParams();
+  const [post, setPost] = useState<Post | null>(null);
+  const [loading, setLoading] = useState(false);
+  const [error, setError] = useState<Error | null>(null);
 
-  //idを数値に変換
-  const numId = Number(id);
-  return <Post id={numId} />;
+  useEffect(() => {
+    const fetcher = async (): Promise<void> => {
+      try {
+        setLoading(true);
+        const res = await fetch(`${API_BASE_URL}/posts/${id}`);
+        if (!res.ok) {
+          throw new Error("データが見つかりません");
+        }
+        const { post } = (await res.json()) as { post: Post };
+        console.log(post);
+        setPost(post);
+      } catch (error) {
+        if (error instanceof Error) {
+          setError(error);
+        }
+      } finally {
+        setLoading(false);
+      }
+    };
+    fetcher();
+  }, [id]);
+
+  if (loading) {
+    return <div>読込中・・・</div>;
+  }
+
+  if (error) {
+    return <div>一覧記事取得エラー: {error.message}</div>;
+  }
+
+  if (!post) {
+    return <div>記事が存在しません</div>;
+  }
+
+  return (
+    <main>
+      <div className="post-detail">
+        <img src={post.thumbnailUrl} alt={post.title} />
+        <div className="post-info">
+          <p>{new Date(post.createdAt).toLocaleDateString()}</p>
+          <ul>
+            {post.categories?.map((category) => (
+              <li key={category}>{category}</li>
+            ))}
+          </ul>
+        </div>
+        <h1>{post.title}</h1>
+        <p dangerouslySetInnerHTML={{ __html: post.content }} />
+      </div>
+    </main>
+  );
 }
